tracker_pnl/requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2

# API clients
requests==2.31.0
aiohttp==3.9.1

# SharePoint integration
Office365-REST-Python-Client==2.5.3
msal==1.25.0

# HTML parsing for Telegram messages
beautifulsoup4==4.12.2
lxml==5.1.0

# Date/time handling
pytz==2023.3

# Logging and utilities
python-dateutil==2.8.2
